==> 2_Meet in the middle.Rmd <==
---
title: 'Today project'
author: "Hongxu Wang"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
<style type="text/css">
body{
font-size: 12pt;
}
</style>

```{r exposure outcome setup, include=FALSE}
source(here::here("!libraries.R"))
source(here::here("!directories.R"))
source(here::here("!load_clean_data.R"))
source(here::here("!functions.R"))

options(knitr.table.format = "html")
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Meet in the middle
* Mediator ~ Exposure
* Outcome ~ Mediator
* covariates: sex, age, kidney function, dxtime
* Result: 35 proteins are significantly associated with both exposures and outcomes
* We will not use the effect estimate from Outcome ~ mediator for plotting
```{r meet in the middle}
# 1) Exposure-Mediator regressions ----------------
result_em <- epiomics::owas(df = data_scaled, 
                            var = "pfas_pfna", #analysis_pfas_names,
                            omics = omic_names, #prot_names,
                            covars = covars,
                            var_exposure_or_outcome = "exposure",
                            family = "gaussian",
                            confidence_level = 0.95, 
                            conf_int = TRUE)

# Calculate significance based on effective number of tests (42)
result_em <- result_em |> 
  mutate(omic_layer = if_else(feature_name %in% prot_names, "Proteomics", "Metabolomics")) |> 
  group_by(var_name, omic_layer) |>
  mutate(sig_efnum = if_else(p_value < (0.05/42), "Sig", "Not sig"), 
         adjusted_pval = p.adjust(p_value, method = "fdr"),
         sig_fdr = if_else(adjusted_pval < 0.2, "Sig", "Not sig")) |> 
  ungroup()

# Check seq.8032.23 (MCCD1)

## A. Clean results -----
# get name of all sig proteins
omic_names_sig <- result_em |> 
  dplyr::filter(p_value < 0.05)
omic_names_sig <- unique(omic_names_sig$feature_name)

# Select/rename key columns
result_em_fin <- result_em %>% 
  dplyr::select(omic_layer, var_name, feature_name, test_statistic,
                estimate, conf_low, conf_high, p_value, adjusted_pval, sig_efnum, sig_fdr) |> 
  rename(estimate_em = estimate, 
         conf_low_em =  conf_low, 
         conf_high_em = conf_high,
         p_value_em = p_value, 
         test_statistic_em = test_statistic) 

# filter to significant omics only
result_em_sig <- result_em_fin %>% 
 tidylog::filter(feature_name %in% omic_names_sig) 


# 2) Mediator-outcome regressions ----------------
ind_vars <- tibble(exposure = omic_names)
dep_vars <- tibble(time = c("daystomic"), 
                   event = c("mic"))

# Get exposure and outcome combinations
mo_comb <- list(omic = ind_vars$exposure, event = dep_vars$event) %>%
  cross_df() %>% 
  left_join(dep_vars, by = "event")

# Create combination of exposures and outcomes
mo_res_models <- mo_comb %>%
  mutate(covar = str_c(covars, collapse = "+"),
         formula = str_c("Surv(", time, ",", event, ")", "~", omic, "+", covar))

# Run all models
mo_res_models$output <- map(mo_res_models$formula,
                            ~coxph(as.formula(.),
                                   data = data_scaled) %>%
                              tidy(., conf.int = TRUE))

# Modify output
mo_res_models <- mo_res_models %>%
  unnest(output) %>%
  filter(grepl("seq", term)|grepl("met",term))%>%
  mutate(HR = exp(estimate),
         exp_conf_high = exp(conf.high),
         exp_conf_low = exp(conf.low),
         sig = ifelse(p.value < 0.05, "Sig.", "Not Sig.")) 

## A. Clean results ----
# Select key columns, filter significant only
mo_res_models <- mo_res_models %>% 
  mutate(sig_efnum_mo = if_else(p.value < (0.05/42), "Sig", "Not sig"),
         adjusted_pval_mo = p.adjust(p.value, method = "fdr"),
         sig_fdr_mo = if_else(adjusted_pval_mo < 0.2, "Sig", "Not sig")) 

# Select/rename key vars
result_mo_fin <- mo_res_models %>%
  dplyr::select(omic, estimate, p.value, 
                estimate, conf.low, conf.high,
                HR, exp_conf_high, exp_conf_low, adjusted_pval_mo, sig_fdr_mo, sig_efnum_mo) |> 
  rename(feature_name = omic, 
         estimate_mo = estimate, 
         conf_low_mo = conf.low, 
         conf_high_mo = conf.high, 
         p_value_mo = p.value,
         HR_mo = HR,
         exp_conf_low_mo = exp_conf_low,
         exp_conf_high_mo = exp_conf_high)

# Filter sig only
result_mo_sig <- mo_res_models |> 
 tidylog::filter(p.value < 0.05)
```



```{r}
# 3) Combine and analyze meet in middle --------------------
# Combine em and mo
mim_res_all <- tidylog::full_join(result_em_fin, result_mo_fin, by = "feature_name")

# Add proteomics metadata
mim_res_all <- mim_res_all |> 
  tidylog::left_join(prot_metadata %>% 
                       dplyr::select(AptName, EntrezGeneSymbol),
                     by = c("feature_name"="AptName")) |> 
  tidylog::select(EntrezGeneSymbol, everything())

# Clean Gene Names
<<<<<<< HEAD
mim_res_all <- mim_res_all |>
  tidylog::mutate(EntrezGeneSymbol = str_remove_all(EntrezGeneSymbol, 
                                      "_([0-9]|[1-9][0-9]|100)"))

# Clean column names
mim_res_all <- mim_res_all |> 
  tidylog::rename(adjusted_pval_em = adjusted_pval, 
                  sig_efnum_em = sig_efnum,
                  sig_fdr_em = sig_fdr)

# Save results
write_csv(mim_res_all, 
          fs::path(dir_results, 
                   "meet_in_middle_res_all_072924.csv"))  

# Filter to significant only
mim_res_sig <- mim_res_all |> 
  tidylog::filter(p_value_mo < 0.05, p_value_em < 0.05)

mim_res_temp <- mim_res_temp |>
  tidylog::mutate(EntrezGeneSymbol = str_remove(EntrezGeneSymbol, "_2") |>
                    str_remove("_4"))

# seq.19570.12 seq.8355.80 pass FDR for PFNA ~ features
write_csv(mim_res_sig, 
          fs::path(dir_results, 
                   "meet_in_middle_res_sig_020924.csv")) 
```


# Mediator-outcome regressions adjusted by PFNA
* Mediators : sig mediators from meet in the middle
```{r mediator outcome reg adjusted by PFNA}
# 2) Mediator-outcome regressions ----------------
ind_vars <- tibble(exposure = omic_names)
dep_vars <- tibble(time = c("daystomic"), event = c("mic"))

# Get exposure and outcome combinations
mo_comb <- list(omic = ind_vars$exposure, event = dep_vars$event) %>%
  cross_df() %>% 
  left_join(dep_vars, by = "event")

# Create combination of exposures and outcomes
mo_res_models <- mo_comb %>%
  mutate(covar = str_c(covars, collapse = "+"),
         formula = str_c("Surv(", time, ",", event, ")", "~", "pfas_pfna", "+", omic, "+", covar))

# Run all models
mo_res_models$output <- map(mo_res_models$formula,
                            ~coxph(as.formula(.),
                                   data = data_scaled) %>%
                              tidy(., conf.int = TRUE))

# Modify output
mo_res_models <- mo_res_models %>%
  unnest(output) %>%
  filter(grepl("seq", term)|grepl("met",term))%>%
  mutate(HR = exp(estimate),
         exp_conf_high = exp(conf.high),
         exp_conf_low = exp(conf.low)) 

## A. Clean results ----
# Select key columns, filter significant features from meet in the middle
result_mo_sub <- mo_res_models %>% 
  tidylog::filter(term %in% mim_res_temp$feature_name) %>%
  dplyr::select(omic, estimate, conf.low, conf.high, p.value, 
                HR, exp_conf_low, exp_conf_high) |> 
  rename(feature_name = omic, 
         estimate_mo = estimate, 
         conf.low_mo = conf.low, 
         conf.high_mo = conf.high,
         p_value_mo = p.value,
         HR_mo = HR,
         exp_conf_low_mo = exp_conf_low,
         exp_conf_high_mo = exp_conf_high)


# 3) Combine and analyze meet in middle --------------------
# Combine em and mo
mim_res <- tidylog::inner_join(result_em_sig, 
                               result_mo_sub, 
                               by = "feature_name")

# examine meet in middle proteins
mim_res1 <- mim_res |> 
  mutate(emxmo = estimate_em*estimate_mo, 
         max_p = if_else(p_value_em < p_value_mo, p_value_mo, p_value_em))

## Combine with proteomics metadata ------
mim_res2 <- mim_res1 %>% 
  tidylog::left_join(prot_metadata, by = c("feature_name" = "AptName")) |> 
  tidylog::filter(Organism == "Human") 

## check if em * mo are all positive
# mim_res2$emxmo

```
