==> 3_Mediation analysis.Rmd <==
---
title: 'Today project'
author: "Hongxu Wang"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
<style type="text/css">
body{
font-size: 12pt;
}
</style>

```{r exposure outcome setup, include=FALSE}
source(here::here("!libraries.R"))
source(here::here("!directories.R"))
source(here::here("!load_clean_data.R"))
source(here::here("!functions.R"))

options(knitr.table.format = "html")
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Mediation
```{r,results='asis'}
# Read in Meet in middle results, from 2_meet in the middle analysis
# mim_res <- read_csv(fs::path(dir_results, "meet_in_middle_res_df_serum_creat_covar.csv")) 
mim_res <- read_csv(fs::path(dir_results, "meet_in_middle_res_sig_020924.csv")) 

# Run Mediation analysis
med_res_df <- data.frame()
for(name in  mim_res$feature_name){
  cat("### Section ", name, "\n")
  d2 <- data_scaled %>% 
    dplyr::select(pfas_pfna, all_of(c(name, covars)), daystomic, mic) %>% 
    as.data.frame() %>%
    rename(mediator = name)
  # set.seed(1111)
  med <- CMAverse::cmest(data = d2, 
                         # model = "gformula",
                         model = "rb",
                         event = "mic",
                         exposure = "pfas_pfna",
                         mediator = "mediator", 
                         outcome = "daystomic",
                         basec = covars,
                         EMint = TRUE,
                         mreg = list("linear"), 
                         yreg = "coxph",
                         # astar = quantile(d2$pfas_pfna, probs = 0.9),
                         # a = quantile(d2$pfas_pfna, probs = 0.1),
                         astar = 1,
                         a = -1,
                         mval = list(0), 
                         # estimation = "imputation",
                         # inference = "bootstrap",
                         # nboot = 2
                         estimation = "paramfunc",
                         inference = "delta"
  )
  print(summary(med))
  med_res_temp <-  med$effect.pe %>% as.data.frame() %>% rename(pe = ".") %>%
    cbind(med$effect.ci.low %>% as.data.frame() %>% rename(ci_low = ".") ) %>%
    cbind(med$effect.ci.high %>% as.data.frame() %>% rename(ci_high = ".")) %>%
    cbind(med$effect.pval %>% as.data.frame() %>% rename(pval = ".")) %>%
    mutate(feature_name = name) %>%
    rownames_to_column("Effect")%>%
    dplyr::select(feature_name, everything())
  
  med_res_df <- rbind(med_res_df,med_res_temp)
  cat("\n")
}

med_res_df1 <- med_res_df %>% 
  tidylog::left_join(prot_metadata %>% 
                       dplyr::select(AptName, EntrezGeneSymbol),
                     by = c("feature_name"="AptName")) 

## total effect significant
med_res_df_te <- med_res_df %>% 
  filter(Effect == "Rte"& pval < 0.05)

med_res_df_sig <- med_res_df %>%
  filter(Effect == "pm" & feature_name %in% 
           med_res_df_te$feature_name)

## Combine pm info with regression result
res_sig <- med_res_df_sig %>% 
  tidylog::left_join(mim_res %>%
                       dplyr::select(EntrezGeneSymbol, 
                                     feature_name:sig_efnum_mo)) %>% 
  dplyr::select(feature_name, EntrezGeneSymbol, everything())
## 
res <- med_res_df %>% 
  tidylog::left_join(mim_res %>%
                       dplyr::select(EntrezGeneSymbol, 
                                     feature_name:sig_efnum_mo)) %>% 
  dplyr::select(feature_name, EntrezGeneSymbol, everything())

write_csv(res, fs::path(dir_results, "med_res_df_020924.csv")) 
```

# Plot
```{r}
# Coefficient plot HW-----
## result in Long format for plotting (HW)
res_l <- res_sig %>%
  dplyr::select(feature_name, EntrezGeneSymbol,
                estimate_em:p_value_em) %>%
  rename_with(.cols = estimate_em:p_value_em,
              .fn = ~str_remove(., "_em") ) %>%
  mutate(type = "PFNA") %>%
  bind_rows(res_sig %>%
              dplyr::select(feature_name, EntrezGeneSymbol,
                            estimate_mo:conf_high_mo, p.value_mo) %>%
              # estimate_mo:adjusted_pval_mo, p.value_mo) %>%
              rename(estimate  = estimate_mo,
                     conf_low  = conf_low_mo,
                     conf_high = conf_high_mo,
                     p_value = p.value_mo) %>%
              # rename(estimate = HR_mo,
              #        conf_low = exp_conf_low_mo,
              #        conf_high = exp_conf_high_mo,
              #        p_value = p.value_mo) %>%
              mutate(type = "MIC")) %>%
  bind_rows(res_sig %>%
              dplyr::select(feature_name, EntrezGeneSymbol,
                            pe:pval) %>%
              rename(estimate = pe,
                     conf_low = ci_low,
                     conf_high = ci_high,
                     p_value = pval) %>%
              mutate(type = "Proportion Mediated")
  )

dummy2 <- data.frame(type = c("PFNA", "MIC"), Z= c(0, 0))

(coef_plot <- res_l %>%
    filter(type %in% c("PFNA", "MIC"))%>%
    ggplot(aes(x = factor(EntrezGeneSymbol,levels = arrange(res_sig, pe)$EntrezGeneSymbol),
               y = estimate)) +
    geom_errorbar(aes(ymin = conf_low ,
                      ymax = conf_high),
                  width = 0) +
    geom_point() +
    # scale_color_manual(values = c("grey60", "black")) +
    coord_flip() +
    # scale_y_continuous(limits = c(-1.75, 3)) +
    facet_grid(~factor(type, levels=c('PFNA','MIC')),
               scales = "free") +
    geom_hline(data = dummy2, aes(yintercept = Z), linetype = 2, color = "grey50") +
    ylab("β (95% CI)              Log HR (95% CI) ") +
    theme(
      axis.title.y = element_blank(),
      # axis.title.x = element_blank(),
      # strip.text.x = element_blank(),
      panel.background = element_rect(fill="white"),
      strip.background = element_rect(fill = "white"),
      legend.position = "none",
      strip.text.y = element_text(angle = 0, hjust = 0)))

(med_plot <- res_l %>%
    filter(!type %in% c("PFNA", "MIC"))%>%
    ggplot(aes(x = factor(EntrezGeneSymbol,levels = arrange(res_sig, pe)$EntrezGeneSymbol),
               y = 100*estimate)) +
    geom_bar(stat = "identity", color = "grey50", fill = "grey50") +
    # scale_color_manual(values = c("grey60", "black")) +
    coord_flip() +
    # scale_y_continuous(limits = c(-1.75, 3)) +
    facet_grid(~type,
               scales = "free") +
    geom_hline(yintercept = 0, linetype = 2, color = "grey20") +
    ylab("Percentage") +
    theme(
      axis.title.y = element_blank(),
      axis.text.y = element_blank(),
      axis.line.y = element_blank(),
      axis.ticks.y = element_blank(),
      # axis.title.x = element_blank(),
      # strip.text.x = element_blank(),
      panel.background = element_rect(fill="white"),
      strip.background = element_rect(fill = "white"),
      legend.position = "none",
      strip.text.y = element_text(angle = 0, hjust = 0)))

#  scale_color_brewer(name = "Direction", palette = "Dark2" ))

p <- grid.arrange(coef_plot, 
                  med_plot, 
                  nrow = 1,
                  widths = c(4,2))
#Save
ggsave(p,filename = here::here(dir_figure, "Fig2 Mediation.jpeg"),
       width = 8, height = 5)
```
