==> 3_Mediation analysis_mediate.Rmd <==
---
title: 'Today project'
author: "Hongxu Wang"
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
<style type="text/css">
body{
font-size: 12pt;
}
</style>

```{r exposure outcome setup, include=FALSE}
source(here::here("!libraries.R"))
source(here::here("!directories.R"))
source(here::here("!load_clean_data.R"))
source(here::here("!functions.R"))

options(knitr.table.format = "html")
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Pairwise Mediation
```{r,results='asis'}
# Read in Meet in middle results, from 2_meet in the middle analysis
# mim_res <- read_csv(fs::path(dir_results, "meet_in_middle_res_df_serum_creat_covar.csv")) 
mim_res <- read_csv(fs::path(dir_results, "meet_in_middle_res_020924.csv")) 

#Loop to conduct pairwise mediation with multiple exposures and mediators
#Loop repeatedly subsets dataset d1 into d2 for individual pairs of exposures and biomarkers

k = 1
for(name in prot_names){
  
  d2<-data_scaled %>% 
    dplyr::select(pfas_pfna, name, all_of(covars), daystomic, mic) %>% as.data.frame() %>%
    rename(mediator = name)
  set.seed(111)

  med<- mediation::mediate(
    data = d2,
    model.m= lm(mediator~pfas_pfna+sex_male+agebase+eGFR+dxtime,
                  data=d2),
    model.y = survreg(Surv(daystomic,mic)~pfas_pfna+mediator+sex_male+agebase+eGFR+dxtime,
                      data=d2),
    treat="pfas_pfna",
    mediator = "mediator")
  model.y = survreg(Surv(daystomic,mic)~pfas_pfna+mediator+sex_male+agebase+eGFR+dxtime,
                    data=d2)
  # med <- CMAverse::cmest(data = d2, 
  #              model = "gformula",
  #              outcome = "mic",
  #              exposure = "pfas_pfna",
  #              mediator = "mediator", 
  #              event = "daystomic",
  #              basec = c("sex_male","agebase","eGFR","dxtime"),
  #              EMint = TRUE,
  #              mreg = list("linear"), 
  #              yreg = "coxph",
  #              astar = 0,
  #              a = 1,
  #              mval = list(1), 
  #              estimation = "imputation",
  #              inference = "bootstrap",
  #              nboot = 10)
  
  
  
  med.results[k,]<-cbind(nobs(model.y),med$d0, med$d0.ci[1], med$d0.ci[2], med$d0.p, med$d1, med$d1.ci[1],
                         med$d1.ci[2], med$d1.p, med$z0, med$z0.ci[1],med$z0.ci[2], med$z0.p, med$z1, 
                         med$z1.ci[1], med$z1.ci[2], med$z1.p, med$n0, med$n0.ci[1], med$n0.ci[2], 
                         med$n0.p, med$n1, med$n1.ci[1], med$n1.ci[2], med$n1.p, med$tau.coef, med$tau.ci[1], 
                         med$tau.ci[2], med$tau.p, med$d.avg, med$d.avg.ci[1], med$d.avg.ci[2], med$d.avg.p,
                         med$z.avg, med$z.avg.ci[1], med$z.avg.ci[2], med$z.avg.p, med$n.avg, med$n.avg.ci[1], 
                         med$n.avg.ci[2], med$n.avg.p)
        # rownames(med.results)[k]<-paste(colnames(d2)[1],name,colnames(d2)[2+ncovars+1],sep='_')
        rownames(med.results)[k] <- name
        print(k)
        print(rownames(med.results)[k])
        k=k+1
      }

## Adding name to the result
med_result <- med.results %>% rownames_to_column("AptName") %>%
  tidylog::left_join(meta_pro %>% dplyr::select(AptName, EntrezGeneSymbol)) %>%
  dplyr::select(AptName, EntrezGeneSymbol, everything())


write_csv(med_result,fs::path(dir_results,'pairwise_mediation_result.csv'))

```

# Plot
```{r plotting}
med_result <- read.csv(fs::path(dir_results,'pairwise_mediation_result.csv')) %>%
  filter(AptName %in% mim_res$feature_name)

res <- mim_res %>% 
  mutate(EntrezGeneSymbol = ifelse(EntrezGeneSymbol == "FGF8_3", "FGF8", EntrezGeneSymbol)) %>%
  tidylog::left_join(med_result %>% 
                       dplyr::select(-AptName), by = "EntrezGeneSymbol") %>%
  dplyr::mutate(
    EntrezGeneSymbol = case_when(EntrezGeneSymbol == "FGF8_3" ~ "FGF8", 
                                 EntrezGeneSymbol == "HSPA1A_2" ~ "HSPA1A",
                                 EntrezGeneSymbol == "CLEC4G_2" ~ "CLEC4G",
                                 EntrezGeneSymbol == "HEY1_2" ~ "HEY1",
                                 EntrezGeneSymbol == "GPNMB_4" ~ "GPNMB",
                                 EntrezGeneSymbol == "RSPO2_1" ~ "RSPO2",
                                 EntrezGeneSymbol == "IL27|EBI3" ~ "IL27",
                                 EntrezGeneSymbol == "GDF11|MSTN" ~ "GDF11", 
                                 TRUE ~ EntrezGeneSymbol))

# Coefficient plot HW-----
## result in Long format for plotting (HW)
res_l <- res %>%
  dplyr::select(feature_name, EntrezGeneSymbol,
                estimate_em:p_value_em) %>%
  rename_with(.cols = estimate_em:p_value_em,
              .fn = ~str_remove(., "_em") ) %>%
  mutate(type = "PFNA") %>%
  bind_rows(res%>%
              dplyr::select(feature_name, EntrezGeneSymbol,
                            estimate_mo:conf_high_mo, p.value_mo) %>%
              # estimate_mo:adjusted_pval_mo, p.value_mo) %>%
              rename(estimate  = estimate_mo,
                     conf_low  = conf_low_mo,
                     conf_high = conf_high_mo,
                     p_value = p.value_mo) %>%
              # rename(estimate = HR_mo,
              #        conf_low = exp_conf_low_mo,
              #        conf_high = exp_conf_high_mo,
              #        p_value = p.value_mo) %>%
              mutate(type = "Albuminuria")) %>%
  bind_rows(res%>%
              dplyr::select(feature_name, EntrezGeneSymbol,
                            PMed.avg:PMed.avg.pval) %>%
              rename(estimate = PMed.avg,
                     conf_low = PMed.avg.lo,
                     conf_high = PMed.avg.hi,
                     p_value = PMed.avg.pval) %>%
              mutate(type = "Proportion Mediated")
  )

dummy2 <- data.frame(type = c("PFNA", "Albuminuria"), Z= c(0, 0))

(coef_plot <- res_l %>%
    filter(type %in% c("PFNA", "Albuminuria"))%>%
    ggplot(aes(x = factor(EntrezGeneSymbol,levels = arrange(res, PMed.avg)$EntrezGeneSymbol),
               y = estimate)) +
    geom_errorbar(aes(ymin = conf_low ,
                      ymax = conf_high),
                  width = 0) +
    geom_point() +
    # scale_color_manual(values = c("grey60", "black")) +
    coord_flip() +
    # scale_y_continuous(limits = c(-1.75, 3)) +
    facet_grid(~factor(type, levels=c('PFNA','Albuminuria')),
               scales = "free") +
    geom_hline(data = dummy2, aes(yintercept = Z), linetype = 2, color = "grey50") +
    ylab("β (95% CI)              Log HR (95% CI) ") +
    theme(
      axis.title.y = element_blank(),
      # axis.title.x = element_blank(),
      # strip.text.x = element_blank(),
      panel.background = element_rect(fill="white"),
      strip.background = element_rect(fill = "white"),
      legend.position = "none",
      strip.text.y = element_text(angle = 0, hjust = 0)))

(med_plot <- res_l %>%
    filter(!type %in% c("PFNA", "Albuminuria"))%>%
    ggplot(aes(x = factor(EntrezGeneSymbol,levels = arrange(res, PMed.avg)$EntrezGeneSymbol),
               y = estimate)) +
    geom_bar(stat = "identity", color = "white", fill = "grey50", size = 0.001) +
    # scale_color_manual(values = c("grey60", "black")) +
    coord_flip() +
    # scale_y_continuous(limits = c(-1.75, 3)) +
    facet_grid(~type,
               scales = "free") +
    geom_hline(yintercept = 0, linetype = 2, color = "grey20") +
    ylab("Percentage") +
    theme(
      axis.title.y = element_blank(),
      axis.text.y = element_blank(),
      axis.line.y = element_blank(),
      axis.ticks.y = element_blank(),
      # axis.title.x = element_blank(),
      # strip.text.x = element_blank(),
      panel.background = element_rect(fill="white"),
      strip.background = element_rect(fill = "white"),
      legend.position = "none",
      strip.text.y = element_text(angle = 0, hjust = 0)))

#  scale_color_brewer(name = "Direction", palette = "Dark2" ))

p <- grid.arrange(coef_plot, 
                  med_plot, 
                  nrow = 1,
                  widths = c(4,2))
#Save
ggsave(p,filename = here::here(dir_figure, "Fig2 Mediation V3.jpeg"),
       width = 8, height = 6)
```

## Supplemental Table
```{r supplemental table}
est <- read_csv(fs::path(dir_results, 
                          "meet_in_middle_res_020924.csv")) %>%
  mutate(`PFNA β[95%CI]` = paste0(round(estimate_em, 2), " (", 
                               round(conf_low_em, 2), ",",
                               round(conf_high_em, 2),")"),
         `PFNA P-Value` = round(p_value_em, 2),
         `Albuminuria β[95%CI]` = paste0(round(estimate_mo, 2), " (", 
                                          round(conf_low_mo, 2), ",",
                                          round(conf_high_mo, 2),")"),
         `Albuminuria P-Value` = round(p.value_mo, 2))
  

med <- read.csv(fs::path(dir_results,'pairwise_mediation_result.csv')) %>%
  filter(AptName %in% mim_res$feature_name)%>%
  mutate()

# med_w <- med %>% filter(Effect %in% c("Rtnde", "Rpnie", "Rte")) %>%
#   dplyr::select(feature_name, EntrezGeneSymbol,Effect, pe, ci_low, ci_high) %>%
#   pivot_wider(
#               names_from = "Effect",
#               values_from = c(pe, ci_low, ci_high)) %>%
#   mutate(`Rtnde[95%CI]` = paste0(round(pe_Rtnde, 2), " [", 
#                                 round(ci_low_Rtnde, 2), ",",
#                                 round(ci_high_Rtnde, 2),"]"),
#          `Rpnie[95%CI]` = paste0(round(pe_Rpnie, 2), " [",
#                                 round(ci_low_Rpnie, 2), ",",
#                                 round(ci_high_Rpnie, 2),"]"),
#          `Total Effect[95%CI]` = paste0(round(pe_Rte, 2), " [", 
#                                 round(ci_low_Rte, 2), ",",
#                                 round(ci_high_Rte, 2),"]"))

med_w <- med %>% 
  dplyr::select(EntrezGeneSymbol,PMed.avg, PMed.avg.lo, PMed.avg.hi,PMed.avg.pval) %>% 
  # mutate(`Proportion Mediated[95%CI]` = paste0(round(PMed.avg, 2), " (",
  #                               round(PMed.avg.lo, 2), ",",
  #                               round(PMed.avg.hi, 2),")"),
  #        `P-value` = PMed.avg.pval)
  mutate(`Proportion Mediated` = paste0(round(PMed.avg, 2)),
         `P-value` = PMed.avg.pval)

df <- est %>%  mutate(EntrezGeneSymbol = ifelse(EntrezGeneSymbol == "FGF8_3", "FGF8", EntrezGeneSymbol)) %>%
  dplyr::select(EntrezGeneSymbol, contains("PFNA"), contains("Albuminuria")) %>%
  tidylog::left_join(med_w %>% 
                       dplyr::select(EntrezGeneSymbol, `Proportion Mediated`, "P-value"), by = "EntrezGeneSymbol") %>%
  dplyr::mutate(
    EntrezGeneSymbol = case_when(EntrezGeneSymbol == "FGF8_3" ~ "FGF8", 
                                 EntrezGeneSymbol == "HSPA1A_2" ~ "HSPA1A",
                                 EntrezGeneSymbol == "CLEC4G_2" ~ "CLEC4G",
                                 EntrezGeneSymbol == "HEY1_2" ~ "HEY1",
                                 EntrezGeneSymbol == "GPNMB_4" ~ "GPNMB",
                                 EntrezGeneSymbol == "RSPO2_1" ~ "RSPO2",
                                 EntrezGeneSymbol == "IL27|EBI3" ~ "IL27",
                                 EntrezGeneSymbol == "GDF11|MSTN" ~ "GDF11", 
                                 TRUE ~ EntrezGeneSymbol)) %>%
   rename(`Protein name` = EntrezGeneSymbol)
  


writexl::write_xlsx(df, fs::path(dir_results, "Supplemental Table 2.xlsx"))
```
